Update code.py
Added optional mode for using accelerometer during muscle activity mode. This feature is disabled on startup. When enabled it will force the leg to straighten when the leg swings forward past a set threshold. Enable and set the threshold by moving slider all the way to the zero-position and use the bend-button in accelerometer mode to walk (this is essentially full manual mode). The program will note the position you first let go of the button to straighten your leg during each step and use it during the muscle activity mode to do the same. To disable this feature, go back to accelerometer mode, move the slider to any position other than zero, and press and release the bend button again. A white flash will indicate when you have enabled and set, or disabled the setting successfully.